Clean up legacy backward compatibility files and update imports
- Remove obsolete session files: base.py, bash.py, tmux.py, subprocess.py, powershell.py
- Update all imports to use new unified terminal architecture
- Fix type annotations in test files to use proper casting
- Update test files to use create_terminal_session factory function
- All tests passing with clean architecture
