+++ openhands/tools/execute_bash/sessions/__init__.py
@@ -1,47 +1,33 @@
-# Terminal session implementations
+# Terminal session implementations - New Unified Architecture
 
-# Base classes and factory
-from openhands.tools.execute_bash.sessions.base import (
-    TerminalCommandStatus,
-    TerminalSession,
-)
-
-# Backward compatibility
-from openhands.tools.execute_bash.sessions.bash import BashCommandStatus, BashSession
+# Factory for creating sessions
 from openhands.tools.execute_bash.sessions.factory import create_terminal_session
-
-# Legacy session implementations (for backward compatibility)
-from openhands.tools.execute_bash.sessions.powershell import PowershellSession
 from openhands.tools.execute_bash.sessions.powershell_terminal import PowershellTerminal
-from openhands.tools.execute_bash.sessions.subprocess import SubprocessBashSession
 from openhands.tools.execute_bash.sessions.subprocess_terminal import SubprocessTerminal
 
 # Terminal interface and backends
 from openhands.tools.execute_bash.sessions.terminal_interface import TerminalInterface
-from openhands.tools.execute_bash.sessions.tmux import TmuxBashSession
 from openhands.tools.execute_bash.sessions.tmux_terminal import TmuxTerminal
 
-# Unified session
-from openhands.tools.execute_bash.sessions.unified_session import UnifiedBashSession
+# Unified session controller and base classes
+from openhands.tools.execute_bash.sessions.unified_session import (
+    TerminalCommandStatus,
+    TerminalSession,
+    UnifiedBashSession,
+)
 
 
 __all__ = [
-    # === Base Classes and Factory ===
+    # === Factory ===
+    "create_terminal_session",
+    # === Base Classes ===
     "TerminalSession",
     "TerminalCommandStatus",
-    "create_terminal_session",
     # === Terminal Interface and Backends ===
     "TerminalInterface",
     "TmuxTerminal",
     "SubprocessTerminal",
     "PowershellTerminal",
-    # === Unified Session ===
+    # === Unified Session Controller ===
     "UnifiedBashSession",
-    # === Legacy Session Implementations ===
-    "TmuxBashSession",
-    "SubprocessBashSession",
-    "PowershellSession",
-    # === Backward Compatibility ===
-    "BashSession",
-    "BashCommandStatus",
 ]
